Add files via upload
Angel VG
Open the main.ipynb file in the your-code directory. There are exercises on uniform, normal, and exponential distributions. In each exercise please read the question carefully and provide your solutions below the question. All the calculations must be performed using Python. The dataset to be used in the exercise has been provided on GitHub here.